finsh the readme and fix some file bugs.
