--- CODEOWNERS
@@ -1,2 +1,0 @@
-# Comment line immediately above ownership line is reserved for related gus information. Please be careful while editing.
-#ECCN:Open Source
