--- cog.yaml
@@ -1,17 +1,0 @@
-build:
-  gpu: true
-  cuda: "11.1"
-  python_version: "3.8"
-  system_packages:
-    - "libgl1-mesa-glx"
-    - "libglib2.0-0"
-  python_packages:
-    - "ipython==7.30.1"
-    - "torchvision==0.11.1"
-    - "torch==1.10.0"
-    - "timm==0.4.12"
-    - "transformers==4.15.0"
-    - "fairscale==0.4.4"
-    - "pycocoevalcap==1.2"
-
-predict: "predict.py:Predictor"
