+++ requirements.txt
@@ -1,3 +1,4 @@
+torch==1.10.1
 timm>=0.4.12
 transformers>=4.15.0
 fairscale>=0.4.4
